Migrate signing code to a separate module
