+++ examples/cases/createorder/create_conditional_order.py
@@ -15,7 +15,7 @@ from x10.models.order import (
     OrderType,
     TimeInForce,
 )
-from x10.perpetual.order_object import OrderConditionalTriggerParam, create_order_object
+from x10.signing.order_object import OrderConditionalTriggerParam, create_order_object
 
 LOGGER = logging.getLogger()
 MARKET_NAME = BTC_USD_MARKET
