+++ examples/cases/createorder/create_limit_order.py
@@ -8,7 +8,7 @@ from examples.utils import (
     get_adjust_price_by_pct,
 )
 from x10.models.order import OrderSide, TimeInForce
-from x10.perpetual.order_object import create_order_object
+from x10.signing.order_object import create_order_object
 
 LOGGER = logging.getLogger()
 MARKET_NAME = BTC_USD_MARKET
