+++ examples/cases/createorder/create_market_order.py
@@ -3,7 +3,7 @@ from asyncio import run
 
 from examples.utils import BTC_USD_MARKET, create_rest_client
 from x10.models.order import OrderSide, OrderType, TimeInForce
-from x10.perpetual.order_object import create_order_object
+from x10.signing.order_object import create_order_object
 from x10.utils.order import get_price_with_slippage
 
 LOGGER = logging.getLogger()
