+++ tests/perpetual/order_object/test_market_order_object.py
@@ -20,7 +20,7 @@ SLIPPAGE = Decimal("0.0075")
 async def test_create_sell_order(mocker: MockerFixture, create_trading_account, create_btc_usd_market):
     mocker.patch("x10.utils.nonce.generate_nonce", return_value=FROZEN_NONCE)
 
-    from x10.perpetual.order_object import create_order_object
+    from x10.signing.order_object import create_order_object
 
     trading_account = create_trading_account()
     btc_usd_market = create_btc_usd_market()
@@ -87,7 +87,7 @@ async def test_create_sell_order(mocker: MockerFixture, create_trading_account, 
 async def test_create_buy_order(mocker: MockerFixture, create_trading_account, create_btc_usd_market):
     mocker.patch("x10.utils.nonce.generate_nonce", return_value=FROZEN_NONCE)
 
-    from x10.perpetual.order_object import create_order_object
+    from x10.signing.order_object import create_order_object
 
     trading_account = create_trading_account()
     btc_usd_market = create_btc_usd_market()
